refactor: reorganize components and sections
