+++ src/components/common/BulletPointCard.tsx
