+++ src/components/common/FactHighlight.tsx
