+++ src/components/common/InfoTable.tsx
